pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "colonyforge"
version = "0.1.0"
description = "ColonyForge - 自律型ソフトウェア組立システム"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [{ name = "ColonyForge Team" }]
keywords = ["ai", "automation", "mcp", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-ulid>=2.2.0",
    "portalocker>=2.8.0",
    "jcs>=0.2.0",
    "httpx>=0.26.0",
    "litellm>=1.40.0",
    "mcp>=1.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "playwright>=1.41.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-benchmark>=4.0.0",
    "pytest-cov>=4.1.0",
    "pytest-rerunfailures>=14.0.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "types-PyYAML>=6.0.0",
    "pyautogui>=0.9.54",
]

vlm = ["anthropic>=0.18.0", "playwright>=1.41.0", "pyautogui>=0.9.54"]

docs = [
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.27.0",
    "mkdocs-static-i18n>=1.2.0",
    "mkdocs-mermaid2-plugin>=1.1.0",
    "mkdocs-gen-files>=0.5.0",
    "mkdocs-literate-nav>=0.6.0",
    "mkdocs-section-index>=0.3.0",
]

[project.scripts]
colonyforge = "colonyforge.cli:main"
colonyforge-vlm = "colonyforge.vlm_tester.server:main"
colonyforge-agent-ui = "colonyforge.agent_ui.server:main"

[project.urls]
Homepage = "https://github.com/k-iijima/ColonyForge"
Documentation = "https://github.com/k-iijima/ColonyForge#readme"
Repository = "https://github.com/k-iijima/ColonyForge"

[tool.hatch.build.targets.wheel]
packages = ["src/colonyforge"]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = [
    "E501",   # line too long - handled by formatter
    "SIM117", # nested with - sometimes more readable for test setup
    "B017",   # pytest.raises(Exception) - intentional for frozen model tests
]

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short -m 'not e2e and not benchmark' --benchmark-disable"
markers = [
    "e2e: marks tests as end-to-end tests (requires Playwright browser or MCP container)",
    "benchmark: marks tests as performance benchmarks (run with --benchmark-enable)",
    "heavyweight: marks tests with excessive API calls (>50 calls) — never run against cloud APIs in CI",
]

[tool.coverage.run]
branch = true
source = ["src/colonyforge"]
omit = ["*/__pycache__/*", "*/tests/*", "*/vlm_tester/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
    "def _setup_handlers",
    "@self.server.list_tools",
    "@self.server.call_tool",
    "async def list_tools",
    "async def call_tool",
    "if __name__ == .__main__.:",
    "def main",
    "async def run",
    "asyncio.run",
    # 後方互換性プロキシメソッド（実装は handlers で完全にテスト済み）
    "return await self.handlers.handle_",
    "return self.handlers._save_capture",
]
fail_under = 96
show_missing = true
